requirements-test.txt
coverage
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
pytest-django
factory-boy==2.5.2

# Additional test requirements go here

# include common requirements
-r requirements.txt
